Pull request 16 broke the code, it could not be build. Fixed that.
The constructor of User was modified but the sub classes did not use the new constructor of User. For this to work, I introduced two constructors to accomodate both needs.
Furthermore, I used the forename and lastname feature for the test users.
Finally, I adjusted code so that input is always processed and the streams are not closed without any information for the client.